// packages/components/src/ui/Signature/Signature.styles.d.ts
export declare const useStyles: () => {
  container: {
    width: '100%';
    display: 'flex';
    flexDirection: 'column';
  };
  pen: {
    color: '#373A3F';
    backgroundColor: '#FFFFFF';
  };
  signatureWeb: {
    borderColor: '#DEE3EC';
    borderWidth: number;
    borderRadius: import('@alveole/theme').Spacing;
  };
  headerSignature: {
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    justifyContent: 'space-between';
  };
  date: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
  };
};
//# sourceMappingURL=Signature.styles.d.ts.map